+++ back-end/src/main/java/com/br/api/dto/login/LoginDTO.java
@@ -5,10 +5,10 @@ import java.time.LocalDateTime;
 import com.br.api.model.Login.TipoUsuario;
 
 public record LoginDTO (
-    int id_login,
+    Long id_login,
     String nome_usuario,
     String senha,
     TipoUsuario tipo_usuario,
-    int id_usuario,
+    Long id_usuario,
     LocalDateTime data_criacao
 ) { }
